UI in seperated class
